requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
pillow==10.2.0
python-dotenv==1.0.0